+++ src/main/java/com/project/backend/domain/todo/converter/TodoConverter.java
@@ -1,5 +1,6 @@
 package com.project.backend.domain.todo.converter;
 
+import com.project.backend.domain.common.plan.enums.MonthlyWeekdayRule;
 import com.project.backend.domain.member.entity.Member;
 import com.project.backend.domain.todo.dto.request.TodoReqDTO;
 import com.project.backend.domain.todo.dto.response.TodoResDTO;
