+++ src/main/java/com/project/backend/domain/todo/entity/TodoRecurrenceGroup.java
@@ -1,7 +1,6 @@
 package com.project.backend.domain.todo.entity;
 
 import com.project.backend.domain.event.enums.MonthlyType;
-import com.project.backend.domain.event.enums.MonthlyWeekdayRule;
 import com.project.backend.domain.event.enums.RecurrenceEndType;
 import com.project.backend.domain.event.enums.RecurrenceFrequency;
 import com.project.backend.domain.member.entity.Member;
