+++ src/main/java/com/project/backend/global/recurrence/util/RecurrenceUtils.java
@@ -1,5 +1,6 @@
 package com.project.backend.global.recurrence.util;
 
+import com.project.backend.domain.common.plan.enums.MonthlyWeekdayRule;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
